# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: scans
    description: Card scan analysis
  - name: prices
    description: Card price observations
  - name: catalog
    description: Curated catalog cards
  - name: supabase
    description: Supabase connection
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /supabase/status:
    get:
      operationId: getSupabaseStatus
      tags: [supabase]
      summary: Check the Supabase project connection
      responses:
        "200":
          description: The project API is reachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupabaseStatus"
        "503":
          description: The project API is unavailable
  /scans/analyze:
    post:
      operationId: analyzeScan
      tags: [scans]
      summary: Analyze a Japanese Pokemon card photo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeScanRequest"
      responses:
        "200":
          description: Card analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardAnalysis"
        "400":
          description: Invalid image request
        "413":
          description: Image is too large
        "502":
          description: Card analysis service unavailable
  /scans/condition:
    post:
      operationId: analyzeCondition
      tags: [scans]
      summary: Assess visible card condition from one or more photos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConditionScanInput"
      responses:
        "200":
          description: Visible condition and image quality assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConditionAnalysis"
        "400":
          description: Invalid image request
        "413":
          description: Image request is too large
        "429":
          description: Too many condition analyses
        "502":
          description: Condition analysis service unavailable
  /cards/{cardId}/prices:
    get:
      operationId: getCardPrices
      tags: [prices]
      summary: Get site-by-site card prices
      parameters:
        - name: cardId
          in: path
          required: true
          description: Legacy printed number or canonical catalog UUID.
          schema: { type: string, minLength: 1, maxLength: 80 }
        - name: period
          in: query
          required: false
          schema: { type: integer, enum: [7, 30, 90, 365], default: 30 }
        - name: demo
          in: query
          required: false
          schema: { type: boolean, default: false }
        - name: name
          in: query
          required: false
          schema: { type: string, maxLength: 100 }
      responses:
        "200":
          description: Site-by-site prices and transaction aggregation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardPrices"
        "400":
          description: Invalid period
        "404":
          description: Catalog UUID was not found
  /cards/featured:
    get:
      operationId: getFeaturedCards
      tags: [catalog]
      summary: Get curated featured cards with confirmed recent sale prices when available
      responses:
        "200":
          description: Curated cards with rights-approved primary images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeaturedCardsResponse"
        "503":
          description: Catalog database unavailable
  /cards/catalog/{cardId}:
    get:
      operationId: getCatalogCard
      tags: [catalog]
      summary: Get a catalog card by canonical UUID
      parameters:
        - name: cardId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Catalog card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogCardResponse"
        "400":
          description: Invalid catalog UUID
        "404":
          description: Catalog card not found
        "503":
          description: Catalog database unavailable
  /cards/market-analysis:
    post:
      operationId: analyzeMarket
      tags: [prices]
      summary: Deterministic card market indicators and optional AI explanation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketAnalysisInput"
      responses:
        "200":
          description: Market indicators with explicit missing-data fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketAnalysis"
        "400":
          description: Invalid card identity or condition snapshot
        "429":
          description: Too many market analyses
        "502":
          description: Price provider unavailable
  /cards/ai-market-search:
    post:
      operationId: searchCardMarketWithAi
      tags: [prices]
      summary: Search Japanese card price references using AI web research
      description: Returns cited current market signals from an OpenAI Responses API web search. AI estimates are references only and are not stored as confirmed price observations.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiMarketSearchRequest"
      responses:
        "200":
          description: Current Japanese market research with cited source signals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiMarketSearchResponse"
        "400":
          description: Card identity is incomplete or invalid
        "429":
          description: Search rate limit exceeded
        "502":
          description: AI web research service unavailable or returned invalid data
components:
  schemas:
    CatalogCard:
      type: object
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        number: { type: string }
        series: { type: string }
        rarity: { type: string }
        imageUrl: { type: [string, "null"], format: uri }
      required: [id, name, number, series, rarity, imageUrl]
    CatalogCardResponse:
      type: object
      additionalProperties: false
      properties:
        card: { $ref: "#/components/schemas/CatalogCard" }
      required: [card]
    FeaturedCard:
      type: object
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        number: { type: string }
        series: { type: string }
        rarity: { type: string }
        imageUrl: { type: string, format: uri }
        marketPrice: { type: [number, "null"], minimum: 0 }
        marketPriceBasis:
          type: [string, "null"]
          enum: [confirmed_ungraded_sales, null]
        transactionCount: { type: integer, minimum: 0 }
      required: [id, name, number, series, rarity, imageUrl, marketPrice, marketPriceBasis, transactionCount]
    FeaturedCardsResponse:
      type: object
      additionalProperties: false
      properties:
        cards:
          type: array
          maxItems: 12
          items: { $ref: "#/components/schemas/FeaturedCard" }
      required: [cards]
    SupabaseStatus:
      type: object
      properties:
        connected: { type: boolean }
      required: [connected]
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AnalyzeScanRequest:
      type: object
      additionalProperties: false
      properties:
        imageBase64:
          type: string
          description: Base64-encoded JPEG or PNG image, no larger than 5 MiB decoded.
        mimeType:
          type: string
          enum: [image/jpeg, image/png]
      required: [imageBase64, mimeType]
    ConditionImage:
      type: object
      additionalProperties: false
      properties:
        imageBase64:
          type: string
          description: Base64-encoded JPEG or PNG image, no larger than 5 MiB decoded.
        mimeType:
          type: string
          enum: [image/jpeg, image/png]
        view:
          type: string
          enum: [front, back, top-left, top-right, bottom-left, bottom-right]
      required: [imageBase64, mimeType, view]
    ConditionScanInput:
      type: object
      additionalProperties: false
      properties:
        images:
          type: array
          description: Up to six photos; 5 MiB decoded in total across all photos.
          minItems: 1
          maxItems: 6
          items: { $ref: "#/components/schemas/ConditionImage" }
      required: [images]
    ConditionJudgement:
      type: object
      additionalProperties: false
      properties:
        status:
          type: string
          enum: [good, minor, moderate, significant, uncertain, not_assessable]
        confidence: { type: number, minimum: 0, maximum: 1 }
        note: { type: string }
        count:
          type: string
          description: Number of visible defects; unknown when the image does not support a count.
          enum: [none, one, few, many, unknown]
      required: [status, confidence, note, count]
    ConditionQualityChecks:
      type: object
      additionalProperties: false
      properties:
        wholeCardVisible: { type: boolean }
        focusSufficient: { type: boolean }
        strongGlare: { type: boolean }
        cropped: { type: boolean }
        conditionAssessable: { type: boolean }
      required: [wholeCardVisible, focusSufficient, strongGlare, cropped, conditionAssessable]
    ConditionAnalysis:
      type: object
      additionalProperties: false
      properties:
        surface: { $ref: "#/components/schemas/ConditionJudgement" }
        corners: { $ref: "#/components/schemas/ConditionJudgement" }
        edges: { $ref: "#/components/schemas/ConditionJudgement" }
        whitening: { $ref: "#/components/schemas/ConditionJudgement" }
        centering: { $ref: "#/components/schemas/ConditionJudgement" }
        scratches: { $ref: "#/components/schemas/ConditionJudgement" }
        dents: { $ref: "#/components/schemas/ConditionJudgement" }
        creases: { $ref: "#/components/schemas/ConditionJudgement" }
        peeling: { $ref: "#/components/schemas/ConditionJudgement" }
        water_damage: { $ref: "#/components/schemas/ConditionJudgement" }
        overall_rank:
          type: string
          description: CARD EYE image-based condition rank, not a professional grading equivalent.
          enum: [S, A, A-, B, C, D, unassessable]
        rank_confidence: { type: number, minimum: 0, maximum: 1 }
        rank_reason: { type: string }
        overallConfidence: { type: number, minimum: 0, maximum: 1 }
        imageQuality:
          type: string
          enum: [acceptable, limited, unusable]
        retakeRecommended: { type: boolean }
        qualityChecks: { $ref: "#/components/schemas/ConditionQualityChecks" }
        limitations:
          type: array
          items: { type: string }
      required: [surface, corners, edges, whitening, centering, scratches, dents, creases, peeling, water_damage, overall_rank, rank_confidence, rank_reason, overallConfidence, imageQuality, retakeRecommended, qualityChecks, limitations]
    CardObservations:
      type: object
      additionalProperties: false
      properties:
        centering: { type: [string, "null"] }
        corners: { type: [string, "null"] }
        edges: { type: [string, "null"] }
        surface: { type: [string, "null"] }
        dirt: { type: [string, "null"] }
        other: { type: [string, "null"] }
      required: [centering, corners, edges, surface, dirt, other]
    CardAnalysis:
      type: object
      additionalProperties: false
      properties:
        identified: { type: boolean }
        cardName: { type: [string, "null"] }
        series: { type: [string, "null"] }
        cardNumber: { type: [string, "null"] }
        rarity: { type: [string, "null"] }
        catalogMatch: { $ref: "#/components/schemas/CatalogMatch" }
        conditionSummary: { type: [string, "null"] }
        observations:
          $ref: "#/components/schemas/CardObservations"
      required: [identified, cardName, series, cardNumber, rarity, conditionSummary, observations]
    CatalogMatch:
      type: object
      additionalProperties: false
      properties:
        status: { type: string, enum: [exact, high, ambiguous, unmatched, unavailable] }
        matchedCardId: { type: [string, "null"], format: uuid }
        method: { type: [string, "null"] }
        candidates:
          type: array
          items: { $ref: "#/components/schemas/CatalogMatchCandidate" }
      required: [status, matchedCardId, method, candidates]
    CatalogMatchCandidate:
      type: object
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        collectorNumber: { type: string }
        setCode: { type: string }
        setName: { type: string }
        rarity: { type: [string, "null"] }
        variantCode: { type: string }
      required: [id, name, collectorNumber, setCode, setName, rarity, variantCode]
    PriceSourceConfig:
      type: object
      additionalProperties: false
      properties:
        source: { type: string }
        displayName: { type: string }
        type: { type: string, enum: [SHOP, MARKETPLACE] }
        capabilities:
          type: array
          items: { type: string, enum: [LISTING, TRANSACTION, BUYBACK] }
      required: [source, displayName, type, capabilities]
    PriceSourceAvailability:
      type: object
      additionalProperties: false
      properties:
        source: { type: string }
        priceType: { type: string, enum: [LISTING, SALE, BUYBACK] }
        status: { type: string, enum: [available, no_data, unavailable] }
        reason: { type: [string, "null"] }
      required: [source, priceType, status, reason]
    PriceListing:
      type: object
      additionalProperties: false
      properties:
        source: { type: string }
        displayName: { type: string }
        priceType: { type: string, enum: [LISTING] }
        price: { type: number }
        isReference: { type: boolean }
        lastUpdated: { type: string, format: date-time }
        condition: { type: [string, "null"] }
        stockStatus: { type: [string, "null"] }
        history:
          type: array
          items: { $ref: "#/components/schemas/PriceHistoryPoint" }
      required: [source, displayName, priceType, price, isReference, lastUpdated, condition, stockStatus, history]
    PriceHistoryPoint:
      type: object
      additionalProperties: false
      properties:
        date: { type: string, format: date-time }
        price: { type: number }
      required: [date, price]
    PriceTransactionSummary:
      type: object
      additionalProperties: false
      properties:
        source: { type: string }
        displayName: { type: string }
        priceType: { type: string, enum: [SALE] }
        medianPrice: { type: [number, "null"] }
        transactionCount: { type: integer }
        highestPrice: { type: [number, "null"] }
        lowestPrice: { type: [number, "null"] }
        lastUpdated: { type: [string, "null"], format: date-time }
        history:
          type: array
          items: { $ref: "#/components/schemas/PriceHistoryPoint" }
      required: [source, displayName, priceType, medianPrice, transactionCount, highestPrice, lowestPrice, lastUpdated, history]
    PriceBuyback:
      type: object
      additionalProperties: false
      properties:
        source: { type: string }
        displayName: { type: string }
        priceType: { type: string, enum: [BUYBACK] }
        price: { type: number }
        lastUpdated: { type: string, format: date-time }
      required: [source, displayName, priceType, price, lastUpdated]
    PriceSummary:
      type: object
      additionalProperties: false
      properties:
        transactionMedian: { type: [number, "null"] }
        shopMedian: { type: [number, "null"] }
        buybackMedian: { type: [number, "null"] }
        psa10Median: { type: [number, "null"] }
        transactionCount: { type: integer }
        confidenceScore: { type: [number, "null"] }
        highestPrice: { type: [number, "null"] }
        lowestPrice: { type: [number, "null"] }
        changePercent: { type: [number, "null"] }
      required: [transactionMedian, shopMedian, buybackMedian, psa10Median, transactionCount, confidenceScore, highestPrice, lowestPrice, changePercent]
    PriceObservation:
      type: object
      additionalProperties: false
      properties:
        source: { type: string }
        sourceType: { type: string, enum: [SHOP, MARKETPLACE] }
        observedAt: { type: string, format: date-time }
        price: { type: number, exclusiveMinimum: 0 }
        condition: { type: [string, "null"] }
        graded: { type: boolean, enum: [false] }
        grade: { type: "null" }
        saleStatus: { type: string, enum: [sold, listing, buyback] }
      required: [source, sourceType, observedAt, price, condition, graded, grade, saleStatus]
    CardPrices:
      type: object
      additionalProperties: false
      properties:
        cardId: { type: string }
        currency: { type: string, enum: [JPY] }
        mode: { type: string, enum: [demo, live] }
        periodDays: { type: integer, enum: [7, 30, 90, 365] }
        marketPrice: { type: [number, "null"] }
        marketPriceConfidence: { type: string, enum: [high, medium, low, insufficient] }
        marketPriceBasis: { type: [string, "null"], enum: [confirmed_ungraded_sales, shop_listing_reference, null] }
        observations:
          type: array
          items: { $ref: "#/components/schemas/PriceObservation" }
        reference:
          oneOf:
            - $ref: "#/components/schemas/PriceReference"
            - type: "null"
        summary: { $ref: "#/components/schemas/PriceSummary" }
        sources:
          type: object
          additionalProperties: false
          properties:
            sales: { type: array, items: { $ref: "#/components/schemas/PriceListing" } }
            transactions: { type: array, items: { $ref: "#/components/schemas/PriceTransactionSummary" } }
            buybacks: { type: array, items: { $ref: "#/components/schemas/PriceBuyback" } }
          required: [sales, transactions, buybacks]
        sourceConfigs:
          type: array
          items: { $ref: "#/components/schemas/PriceSourceConfig" }
        sourceAvailability:
          type: array
          items: { $ref: "#/components/schemas/PriceSourceAvailability" }
        methodology: { type: string }
      required: [cardId, currency, mode, periodDays, marketPrice, marketPriceConfidence, marketPriceBasis, observations, reference, summary, sources, sourceConfigs, methodology]
    PriceReference:
      type: object
      additionalProperties: false
      properties:
        source: { type: string }
        price: { type: number }
      required: [source, price]
    MarketAnalysisInput:
      type: object
      additionalProperties: false
      properties:
        cardId: { type: string, minLength: 1, maxLength: 40 }
        name: { type: string, minLength: 1, maxLength: 100 }
        condition:
          oneOf:
            - $ref: "#/components/schemas/ConditionAnalysis"
            - type: "null"
      required: [cardId, name, condition]
    AiMarketSearchRequest:
      type: object
      additionalProperties: false
      properties:
        cardName: { type: string, minLength: 1, maxLength: 100 }
        cardNumber: { type: [string, "null"], maxLength: 40 }
        series: { type: [string, "null"], maxLength: 100 }
        rarity: { type: [string, "null"], maxLength: 40 }
      required: [cardName]
    AiMarketSearchSource:
      type: object
      additionalProperties: false
      properties:
        title: { type: string }
        url: { type: string, format: uri }
        category: { type: string, enum: [sale, shop, buyback, psa10] }
        price: { type: [number, "null"], exclusiveMinimum: 0 }
      required: [title, url, category, price]
    AiMarketSearchResponse:
      type: object
      additionalProperties: false
      properties:
        cardName: { type: string }
        cardNumber: { type: [string, "null"] }
        series: { type: [string, "null"] }
        rarity: { type: [string, "null"] }
        searchedAt: { type: string, format: date-time }
        marketPrice: { type: [number, "null"], exclusiveMinimum: 0 }
        saleMedian: { type: [number, "null"], exclusiveMinimum: 0 }
        saleCount: { type: [integer, "null"], minimum: 1 }
        shopMin: { type: [number, "null"], exclusiveMinimum: 0 }
        shopMax: { type: [number, "null"], exclusiveMinimum: 0 }
        buybackMin: { type: [number, "null"], exclusiveMinimum: 0 }
        buybackMax: { type: [number, "null"], exclusiveMinimum: 0 }
        psa10Median: { type: [number, "null"], exclusiveMinimum: 0 }
        explanation: { type: string }
        sources:
          type: array
          items: { $ref: "#/components/schemas/AiMarketSearchSource" }
      required: [cardName, cardNumber, series, rarity, searchedAt, marketPrice, saleMedian, saleCount, shopMin, shopMax, buybackMin, buybackMax, psa10Median, explanation, sources]
    MarketEvent:
      type: object
      additionalProperties: false
      properties:
        title: { type: string }
        source: { type: string }
        observedAt: { type: string, format: date-time }
      required: [title, source, observedAt]
    MarketIndicators:
      type: object
      additionalProperties: false
      properties:
        priceLevel: { type: string }
        shortTermTrend: { type: string }
        volatility: { type: string }
        activity: { type: string }
        condition: { type: string }
        marketInfo: { type: string }
        reliability: { type: string }
      required: [priceLevel, shortTermTrend, volatility, activity, condition, marketInfo, reliability]
    MarketAnalysis:
      type: object
      additionalProperties: false
      properties:
        cardId: { type: string }
        currency: { type: string, enum: [JPY] }
        currentPrice: { type: [number, "null"] }
        currentPriceBasis: { type: ["string", "null"], enum: [SALE, LISTING, null] }
        average7d: { type: [number, "null"] }
        average30d: { type: [number, "null"] }
        average90d: { type: [number, "null"] }
        change7d: { type: [number, "null"] }
        change30d: { type: [number, "null"] }
        deviationFrom90d: { type: [number, "null"] }
        trend: { type: string, enum: [rising, falling, stable, insufficient] }
        volatilityPercent: { type: [number, "null"] }
        volume7d: { type: integer }
        volume30d: { type: integer }
        volume90d: { type: integer }
        volumeChangePercent: { type: [number, "null"] }
        conditionSummary: { type: [string, "null"] }
        marketEvents: { type: array, items: { $ref: "#/components/schemas/MarketEvent" } }
        dataConfidence: { type: string, enum: [high, medium, low, insufficient] }
        confidenceReasons: { type: array, items: { type: string } }
        indicators: { $ref: "#/components/schemas/MarketIndicators" }
        analysisSummary: { type: string }
        aiExplanation: { type: [string, "null"] }
        calculatedAt: { type: string, format: date-time }
        dataAsOf: { type: [string, "null"], format: date-time }
        dataSources: { type: array, items: { type: string } }
        sourceAvailability: { type: array, items: { $ref: "#/components/schemas/PriceSourceAvailability" } }
        history: { type: array, items: { $ref: "#/components/schemas/PriceHistoryPoint" } }
        methodology: { type: string }
      required: [cardId, currency, currentPrice, currentPriceBasis, average7d, average30d, average90d, change7d, change30d, deviationFrom90d, trend, volatilityPercent, volume7d, volume30d, volume90d, volumeChangePercent, conditionSummary, marketEvents, dataConfidence, confidenceReasons, indicators, analysisSummary, aiExplanation, calculatedAt, dataAsOf, dataSources, sourceAvailability, history, methodology]